Make layman package, add username to URL
