+++ .env.dev
@@ -29,7 +29,7 @@ LAYMAN_GS_ROLE=LAYMAN_ROLE
 ##############################################################################
 
 # See http://flask.pocoo.org/docs/1.0/quickstart/
-FLASK_APP=src/layman/layman.py
+FLASK_APP=src/layman
 FLASK_ENV=development
 
 # Flask secret_key, see http://flask.pocoo.org/docs/1.0/quickstart/#sessions
