+++ src/clear-python-cache.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+find . | grep -E "(__pycache__|\.pyc|\.pyo$)" | xargs rm -rf
