+++ start.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
 
-find . | grep -E "(__pycache__|\.pyc|\.pyo$)" | xargs rm -rf
+bash src/clear-python-cache.sh
 python3 src/prepare_layman.py && flask run --host=0.0.0.0 --port=8000
